Correciones y metodo de instancias
cambie el valor que retorna iva, de 5 a 0,05 y así y cree le metodo de instancias

+++ ProyFactura/src/main/java/modelo/Producto.java
@@ -60,7 +60,7 @@ public abstract class Producto {
         this.cantidad = cantidad;
     }
     
-    public abstract int iva();
+    public abstract double iva();
 
     public double valor()
     {
